cleanup
remove redundant functions
